requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.1
fastapi==0.123.7
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
idna==3.11
protobuf==6.33.1
pydantic==2.10.6
pydantic-core==2.27.2
starlette==0.49.3
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
